reverted the CodegenIntegration usage
