+++ examples/energy-management-app/energy-management-common/common/src/EnergyManagementAppCommonMain.cpp
@@ -20,13 +20,13 @@
 #include "EnergyManagementAppCmdLineOptions.h"
 #include <DeviceEnergyManagementDelegateImpl.h>
 #include <DeviceEnergyManagementManager.h>
-#include <ElectricalEnergyMeasurementCluster.h>
 #include <ElectricalPowerMeasurementDelegate.h>
 #include <EnergyEvseMain.h>
 #include <PowerTopologyDelegate.h>
 #include <WhmMain.h>
 #include <WhmManufacturer.h>
 #include <app/clusters/electrical-energy-measurement-server/electrical-energy-measurement-server.h>
+#include <app/clusters/electrical-energy-measurement-server/ElectricalEnergyMeasurementCluster.h>
 #include <device-energy-management-modes.h>
 #include <energy-evse-modes.h>
 #include <water-heater-mode.h>
