+++ examples/energy-management-app/silabs/BUILD.gn
@@ -189,7 +189,6 @@ silabs_executable("energy-management-app") {
   deps = [
     ":sdk",
     "${chip_root}/src/app:attribute-persistence",
-    "${chip_root}/src/app/clusters/electrical-energy-measurement-server:electrical-energy-measurement-server",
     app_data_model,
   ]
