+++ src/app/clusters/electrical-energy-measurement-server/ElectricalEnergyMeasurementCluster.cpp
@@ -15,15 +15,18 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
+#include "electrical-energy-measurement-server.h"
 #include "ElectricalEnergyMeasurementCluster.h"
 
 #include <protocols/interaction_model/StatusCode.h>
 
 #include <app/AttributeAccessInterfaceRegistry.h>
 #include <app/AttributeValueEncoder.h>
-
-#include <app-common/zap-generated/cluster-enums.h>
-#include <app-common/zap-generated/ids/Attributes.h>
+#include <app/EventLogging.h>
+#include <app/reporting/reporting.h>
+#include <app/util/attribute-storage.h>
+#include <app/util/config.h>
+#include <zap-generated/gen_config.h>
 
 using chip::Protocols::InteractionModel::Status;
 
@@ -35,6 +38,9 @@ namespace ElectricalEnergyMeasurement {
 using namespace chip;
 using namespace chip::app::Clusters::ElectricalEnergyMeasurement::Attributes;
 using namespace chip::app::Clusters::ElectricalEnergyMeasurement::Structs;
+
+MeasurementData gMeasurements[MATTER_DM_ELECTRICAL_ENERGY_MEASUREMENT_CLUSTER_SERVER_ENDPOINT_COUNT +
+                              CHIP_DEVICE_CONFIG_DYNAMIC_ENDPOINT_COUNT];
 
 CHIP_ERROR ElectricalEnergyMeasurementAttrAccess::Init()
 {
@@ -136,6 +142,108 @@ bool ElectricalEnergyMeasurementAttrAccess::HasFeature(Feature aFeature) const
 bool ElectricalEnergyMeasurementAttrAccess::SupportsOptAttr(OptionalAttributes aOptionalAttrs) const
 {
     return mOptionalAttrs.Has(aOptionalAttrs);
+}
+
+MeasurementData * MeasurementDataForEndpoint(EndpointId endpointId)
+{
+    auto index = emberAfGetClusterServerEndpointIndex(endpointId, app::Clusters::ElectricalEnergyMeasurement::Id,
+                                                      MATTER_DM_ELECTRICAL_ENERGY_MEASUREMENT_CLUSTER_SERVER_ENDPOINT_COUNT);
+
+    if (index == kEmberInvalidEndpointIndex)
+    {
+        return nullptr;
+    }
+
+    if (index >= MATTER_ARRAY_SIZE(gMeasurements))
+    {
+        ChipLogError(NotSpecified, "Internal error: invalid/unexpected energy measurement index.");
+        return nullptr;
+    }
+    return &gMeasurements[index];
+}
+
+CHIP_ERROR SetMeasurementAccuracy(EndpointId endpointId, const MeasurementAccuracyStruct::Type & accuracy)
+{
+
+    MeasurementData * data = MeasurementDataForEndpoint(endpointId);
+    VerifyOrReturnError(data != nullptr, CHIP_ERROR_INVALID_ARGUMENT);
+
+    data->measurementAccuracy = accuracy;
+
+    MatterReportingAttributeChangeCallback(endpointId, ElectricalEnergyMeasurement::Id, Accuracy::Id);
+
+    return CHIP_NO_ERROR;
+}
+
+CHIP_ERROR SetCumulativeReset(EndpointId endpointId, const Optional<CumulativeEnergyResetStruct::Type> & cumulativeReset)
+{
+
+    MeasurementData * data = MeasurementDataForEndpoint(endpointId);
+    VerifyOrReturnError(data != nullptr, CHIP_ERROR_INVALID_ARGUMENT);
+
+    data->cumulativeReset = cumulativeReset;
+
+    MatterReportingAttributeChangeCallback(endpointId, ElectricalEnergyMeasurement::Id, CumulativeEnergyReset::Id);
+
+    return CHIP_NO_ERROR;
+}
+
+bool NotifyCumulativeEnergyMeasured(EndpointId endpointId, const Optional<EnergyMeasurementStruct::Type> & energyImported,
+                                    const Optional<EnergyMeasurementStruct::Type> & energyExported)
+{
+    MeasurementData * data = MeasurementDataForEndpoint(endpointId);
+    if (data != nullptr)
+    {
+        data->cumulativeImported = energyImported;
+        data->cumulativeExported = energyExported;
+    }
+
+    Events::CumulativeEnergyMeasured::Type event;
+
+    event.energyImported = energyImported;
+    event.energyExported = energyExported;
+
+    EventNumber eventNumber;
+    CHIP_ERROR error = app::LogEvent(event, endpointId, eventNumber);
+
+    if (CHIP_NO_ERROR != error)
+    {
+        ChipLogError(Zcl, "[NotifyCumulativeEnergyMeasured] Unable to send event: %" CHIP_ERROR_FORMAT " [endpointId=%d]",
+                     error.Format(), endpointId);
+        return false;
+    }
+    ChipLogProgress(Zcl, "[NotifyCumulativeEnergyMeasured] Sent event [endpointId=%d,eventNumber=%lu]", endpointId,
+                    static_cast<unsigned long>(eventNumber));
+    return true;
+}
+
+bool NotifyPeriodicEnergyMeasured(EndpointId endpointId, const Optional<EnergyMeasurementStruct::Type> & energyImported,
+                                  const Optional<EnergyMeasurementStruct::Type> & energyExported)
+{
+    MeasurementData * data = MeasurementDataForEndpoint(endpointId);
+    if (data != nullptr)
+    {
+        data->periodicImported = energyImported;
+        data->periodicExported = energyExported;
+    }
+
+    Events::PeriodicEnergyMeasured::Type event;
+
+    event.energyImported = energyImported;
+    event.energyExported = energyExported;
+
+    EventNumber eventNumber;
+    CHIP_ERROR error = app::LogEvent(event, endpointId, eventNumber);
+
+    if (CHIP_NO_ERROR != error)
+    {
+        ChipLogError(Zcl, "[NotifyPeriodicEnergyMeasured] Unable to send event: %" CHIP_ERROR_FORMAT " [endpointId=%d]",
+                     error.Format(), endpointId);
+        return false;
+    }
+    ChipLogProgress(Zcl, "[NotifyPeriodicEnergyMeasured] Sent event [endpointId=%d,eventNumber=%lu]", endpointId,
+                    static_cast<unsigned long>(eventNumber));
+    return true;
 }
 
 } // namespace ElectricalEnergyMeasurement
