+++ src/app/clusters/electrical-energy-measurement-server/app_config_dependent_sources.gni
@@ -12,8 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 app_config_dependent_sources = [
-  "CodegenIntegration.cpp",
-  "CodegenIntegration.h",
+  "ElectricalEnergyMeasurementCluster.cpp",
+  "ElectricalEnergyMeasurementCluster.h",
   "EnergyReportingTestEventTriggerHandler.h",
   "electrical-energy-measurement-server.h",
 ]
